Add files via upload
Shapes Interface:

The Shapes interface serves as a blueprint for defining common behavior applicable to all shapes. It declares two abstract methods: calculateArea() for computing the area of a shape, and calculatePerimeter() for determining its perimeter.

Circle Class:

The Circle class represents a circular shape and implements the Shapes interface. It contains a field to store the radius of the circle. Methods within this class include calculateArea(), which computes the area of the circle using the formula πr^2, and calculatePerimeter(), which calculates the circumference using the formula 2πr. The constructor initializes the radius of the circle.

Triangle Class:

The Triangle class models a triangular shape and implements the Shapes interface. It holds fields for the lengths of its three sides. This class provides methods such as calculateArea(), which employs Heron's formula to determine the area of the triangle, and calculatePerimeter(), which sums the lengths of its sides to find the perimeter. The constructor initializes the lengths of the triangle's sides.

Rectangle Class:

Representing a rectangular shape, the Rectangle class also adheres to the Shapes interface. It maintains fields for the length and width of the rectangle. Methods like calculateArea() compute the area by multiplying its length and width, while calculatePerimeter() determines the perimeter using the formula 2*(length + width). The constructor initializes the length and width of the rectangle.

The Main class is the program's entry point, allowing users to input values for different shapes interactively. It reads user input using Scanner, creates instances of Circle, Triangle, and Rectangle classes with the provided input, and calculates their respective area and perimeter. Finally, it gracefully closes the Scanner. This interactive approach enables users to explore geometric properties dynamically.