Add files via upload
Employee class:
The Employee class is an abstract class representing basic information and functionalities of an employee. It contains attributes such as employee name, address, designation, department, date of joining, bank details, and various allowances and deductions related to salary. It also includes methods for getting and setting employee details, as well as an abstract method salaryCalculation() which needs to be implemented by subclasses.

NormalEmployee class:
The NormalEmployee class extends the Employee class and represents a type of employee with a specific salary calculation method. It implements the salaryCalculation() method to calculate the salary based on a predefined formula. The class includes attributes and methods to perform these calculations, as well as methods to set and get employee details.

BonusEmployee class:
The BonusEmployee class also extends the Employee class and represents another type of employee with a different salary calculation method. It implements the salaryCalculation() method to calculate the salary, incorporating additional calculations for EPF and total deductions. The class includes methods to perform these calculations, as well as methods to set and get employee details.

Main class:
The Main class contains the main method which serves as the entry point for the program. It creates instances of NormalEmployee and BonusEmployee classes, prompts the user to input employee details, and calculates their respective salaries using the methods provided in each class. This class coordinates the execution of the program and interaction with the user.